YOLO_Detection/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
opencv-python>=4.5.0
Pillow>=8.0.0
requests>=2.23.0
PyYAML>=5.3.1
matplotlib>=3.2.2
tqdm>=4.41.0
scikit-learn>=0.22.2
tensorboard>=2.4.1
ultralytics>=8.0.0
